requirements.txt
wheel
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Django==4.0.1
django-extensions==3.1.5
django-filter==21.1
django-rest==0.8.7
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2021.3
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
json2html==1.3.0
